bot/requirements.txt
python-binance==1.0.19
requests==2.31.0
python-dotenv==1.0.0
pandas==2.0.3 